pyproject.toml
[project]
name = "mirage-brandextract-mcp"
version = "0.1.0"
description = "MCP server to extract brand identity from websites and recreate it programmatically"
readme = "README.md"
requires-python = ">=3.11"
license = {text = "MIT"}
authors = [
    {name = "Me, Myself Plus AI LLC"}
]
keywords = ["mcp", "brand", "extraction", "firecrawl", "gemini"]

dependencies = [
    "mcp[cli]>=1.0.0",
    "httpx>=0.26.0",
    "pydantic>=2.0.0",
    "python-dotenv>=1.0.0",
    "google-generativeai>=0.8.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "pytest-asyncio>=0.21.0",
    "black>=23.0",
    "ruff>=0.1.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/mirage"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.ruff]
line-length = 100
target-version = "py311"
